Refactored according to spec

+++ src/main/java/com/booleanuk/api/cinema/controller/CustomerController.java
@@ -1,7 +1,15 @@
 package com.booleanuk.api.cinema.controller;
 
 import com.booleanuk.api.cinema.model.Customer;
+import com.booleanuk.api.cinema.model.Movie;
+import com.booleanuk.api.cinema.model.Screening;
+import com.booleanuk.api.cinema.model.Ticket;
 import com.booleanuk.api.cinema.repository.CustomerRepository;
+import com.booleanuk.api.cinema.repository.TicketRepository;
+import com.booleanuk.api.cinema.response.BadRequestResponse;
+import com.booleanuk.api.cinema.response.MovieResponse;
+import com.booleanuk.api.cinema.response.NotFoundResponse;
+import com.booleanuk.api.cinema.response.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,51 +24,69 @@ public class CustomerController {
 
     @Autowired
     private CustomerRepository customerRepository;
+    @Autowired
+    private TicketRepository ticketRepository;
 
     @GetMapping
-    public List<Customer> getAll() {
-        return this.customerRepository.findAll();
+    public ResponseEntity<Response> getAll() {
+        return ResponseEntity.status(HttpStatus.OK).body(new Response(this.customerRepository.findAll(), "success"));
+
     }
 
     @PostMapping
-    public ResponseEntity<Customer> createCustomer(@RequestBody Customer customer) {
+    public ResponseEntity<Response> createCustomer(@RequestBody Customer customer) {
         if(containsNull(customer)) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Could not create customer, please check all required fields are correct.");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new BadRequestResponse());
         }
-        return new ResponseEntity<>(customerRepository.save(customer), HttpStatus.CREATED);
+        Response response = new Response(customerRepository.save(customer), "success");
+        return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
-
-    @GetMapping("/{id}")
-    public ResponseEntity<Customer> getCustomerById(@PathVariable int id) {
-        Customer customer = findCustomer(id);
-        return ResponseEntity.ok(customer);
-    }
+//    @GetMapping("/{id}")
+//    public ResponseEntity<Customer> getCustomerById(@PathVariable int id) {
+//        Customer customer = findCustomer(id);
+//        return ResponseEntity.ok(customer);
+//    }
 
     //TODO: make sure tickets are deleted
     @DeleteMapping("/{id}")
-    public ResponseEntity<Customer> deleteCustomer(@PathVariable int id) {
+    public ResponseEntity<Response> deleteCustomer(@PathVariable int id) {
         Customer customerToDelete = findCustomer(id);
+        if(customerToDelete == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new NotFoundResponse());
+        }
+
+        for(Ticket ticket: customerToDelete.getTickets()) {
+            ticketRepository.delete(ticket);
+        }
 
         customerRepository.delete(customerToDelete);
-        return ResponseEntity.ok(customerToDelete);
+        return ResponseEntity.status(HttpStatus.OK).body(new Response(customerToDelete, "success"));
     }
 
-    @PutMapping("/{id}")
-    public ResponseEntity<Customer> updateCustomer(@PathVariable int id, @RequestBody Customer customer) {
-        Customer customerToUpdate = findCustomer(id);
-        if(containsNull(customer)) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Could not update the customer, please check all required fields are correct.");
-        }
-        customerToUpdate.setName(customer.getName());
-        customerToUpdate.setPhone(customer.getPhone());
-        customerToUpdate.setEmail(customer.getEmail());
 
-        return new ResponseEntity<>(customerRepository.save(customerToUpdate), HttpStatus.CREATED);
+    @PutMapping("/{id}")
+    public ResponseEntity<Response> updateCustomer(@PathVariable int id, @RequestBody Customer customer) {
+        Customer customerToUpdate = findCustomer(id);
+        if(customerToUpdate == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new NotFoundResponse());
+        }
+
+        if(customer.getName() != null) {
+            customerToUpdate.setName(customer.getName());
+        }
+        if(customer.getPhone() != null) {
+            customerToUpdate.setPhone(customer.getPhone());
+        }
+        if(customer.getEmail() != null) {
+            customerToUpdate.setEmail(customer.getEmail());
+        }
+
+        return ResponseEntity.status(HttpStatus.OK).body(new Response(customerToUpdate, "success"));
     }
 
     private Customer findCustomer(int id) {
-        return customerRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "No customers with that id were found"));
+        return customerRepository.findById(id).orElse(null);
     }
 
     private boolean containsNull(Customer customer) {
